+++ GUI/GUITextPanel.cpp
@@ -487,7 +487,7 @@ void GUITextPanel::UpdateText(bool Typing, bool DoIncrement)
     }
 
     // Clamp it
-    m_StartIndex = std::min(m_StartIndex, m_Text.size()-1);
+    m_StartIndex = std::min(m_StartIndex, static_cast<int>(m_Text.size()-1));
 
     // Adjust the cursor position
     m_CursorX = m_Font->CalculateWidth(m_Text.substr(m_StartIndex, m_CursorIndex-m_StartIndex));
